rebase and update part of the tests
