+++ frontend/src/__tests__/integration/pages/projects/ProjectDetails.spec.ts
@@ -9,8 +9,8 @@ test('Empty project', async ({ page }) => {
   await page.waitForSelector('text=No model servers');
 
   // the dividers number should always 1 less than the section number
-  const sections = await page.locator('[data-testid="details-page-section"]').all();
-  const dividers = await page.locator('[data-testid="details-page-section-divider"]').all();
+  const sections = await page.locator('[data-id="details-page-section"]').all();
+  const dividers = await page.locator('[data-id="details-page-section-divider"]').all();
 
   expect(dividers).toHaveLength(sections.length - 1);
 });
@@ -24,7 +24,7 @@ test('Non-empty project', async ({ page }) => {
   await page.waitForSelector('text=Test Notebook');
 
   // we fill in the page with data, so there should be no dividers on the page
-  expect(await page.locator('[data-testid="details-page-section-divider"]').all()).toHaveLength(0);
+  expect(await page.locator('[data-id="details-page-section-divider"]').all()).toHaveLength(0);
 
   // check the x-small size shown correctly
   expect(await page.getByText('XSmall')).toBeTruthy();
